requirements.txt
python-dotenv
PyMuPDF
openai
requests